+++ app/src/main/java/com/cuidadoanimal/petcare/database/entities/Tag.kt
@@ -5,12 +5,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "Tag")
-data class Tag (
+data class Tag(
     @ColumnInfo(name = "Name")
     val name: String
-)
-
-{
+) {
     @PrimaryKey(autoGenerate = true)
-    var idTag:Int=0
+    var idTag: Long = 0
 }
